ml-microservice/requirements.txt
pandas==2.2.2
scikit-learn==1.5.0
pytest==8.3.2
matplotlib==3.9
notebook==7
ipykernel==6.29.5
xgboost==2.1.1
lightgbm==4.5.0
seaborn==0.13.2
python-dotenv